Reworked filters and modifiers
Filters and Modifiers are now classes with a __call__ method. This
allows for creating the objects with parameter and then calling them
with only the bounding box objects, which is better for chaining.

Moved filters and modifiers to util, because they are not solely for statistics and thus belong in the util package